+++ output/src/main/java/com/ibm/output/behavior/CryptoBehaviorMapper.java
@@ -17,7 +17,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ibm.output.cyclondx.behavior;
+package com.ibm.output.behavior;
 
 import com.ibm.mapper.model.Algorithm;
 import com.ibm.mapper.model.AuthenticatedEncryption;
