+++ output/src/main/java/com/ibm/output/cyclondx/CBOMOutputFile.java
@@ -59,9 +59,9 @@ import com.ibm.mapper.model.protocol.TLS;
 import com.ibm.mapper.utils.DetectionLocation;
 import com.ibm.output.Constants;
 import com.ibm.output.IOutputFile;
+import com.ibm.output.behavior.CryptoBehavior;
+import com.ibm.output.behavior.CryptoBehaviorMapper;
 import com.ibm.output.cyclondx.behavior.BehaviorInferenceEngine;
-import com.ibm.output.cyclondx.behavior.CryptoBehavior;
-import com.ibm.output.cyclondx.behavior.CryptoBehaviorMapper;
 import com.ibm.output.cyclondx.builder.AlgorithmComponentBuilder;
 import com.ibm.output.cyclondx.builder.ProtocolComponentBuilder;
 import com.ibm.output.cyclondx.builder.RelatedCryptoMaterialComponentBuilder;
