+++ output/src/test/java/com/ibm/output/behavior/CryptoBehaviorTaxonomyTest.java
@@ -17,13 +17,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.ibm.output.cyclonedx.behavior;
+package com.ibm.output.behavior;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ibm.output.cyclondx.behavior.CryptoBehavior;
 import java.io.InputStream;
 import java.util.HashSet;
 import java.util.Set;
