+++ output/src/test/java/com/ibm/output/cyclonedx/CryptoBehaviorMetadataTest.java
@@ -35,8 +35,8 @@ import com.ibm.mapper.model.functionality.Encrypt;
 import com.ibm.mapper.model.mode.CBC;
 import com.ibm.mapper.model.protocol.TLS;
 import com.ibm.mapper.utils.DetectionLocation;
+import com.ibm.output.behavior.CryptoBehaviorMapper;
 import com.ibm.output.cyclondx.CBOMOutputFile;
-import com.ibm.output.cyclondx.behavior.CryptoBehaviorMapper;
 import java.util.Collections;
 import java.util.List;
 import org.cyclonedx.model.Bom;
